refactor(config,core): address review comments on v5/v6 struct layout and reactor retry
- Move shared config structs (Connection, WireGuard, BlokliConfig and all
  helpers/From-impls) to v6 (canonical module); v5 now re-exports them via
  pub(super) use super::v6::...
- Add validate_hops on v6 DestinationPath::Hops, exposed pub(super) for v5
- Replace spawn_ticket_stats_runner fallback in try_start_reactor with a
  dedicated spawn_retry_reactor that sends Results::RetryReactor, keeping
  reactor retry independent of the ticket-stats polling cadence
- Add Results::RetryReactor variant and handle it in on_results
